+++ src/main/java/com/albee/mydataapi/api/base/tele/models/entity/TeleHistEntity.java
@@ -13,6 +13,6 @@ import java.math.BigDecimal;
 public class TeleHistEntity extends EntityBase {
     private String mgmtId;    // 계약관리번호
     private String transMonth;    // 납부연월
-    private BigDecimal paidAmt;    // 납부금액
+    private Long paidAmt;    // 납부금액
     private String payMethod;    // 납부수단(코드)
 }
