+++ src/main/java/com/albee/mydataapi/api/base/tele/models/form/TeleChargeForm.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class TeleChargeForm extends FormBase {
     private String mgmtId;    // 계약관리번호
     private String chargeMonth;    // 청구년월
-    private BigDecimal chargeAmt;    // 청구금액
+    private Long chargeAmt;    // 청구금액
     private String chargeDate;    // 납부예정일자
 
     public TeleChargeEntity getEntity(){
