+++ src/main/java/com/albee/mydataapi/api/base/tele/models/form/TeleHistForm.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 public class TeleHistForm extends FormBase {
     private String mgmtId;    // 계약관리번호
     private String transMonth;    // 납부연월
-    private BigDecimal paidAmt;    // 납부금액
+    private Long paidAmt;    // 납부금액
     private String payMethod;    // 납부수단(코드)
 
     public TeleHistEntity getEntity(){
